+++ packages/core-js/actual/string/virtual/index.js
@@ -1,3 +1,6 @@
 var parent = require('../../../stable/string/virtual');
 
+require('../../../modules/esnext.string.is-well-formed');
+require('../../../modules/esnext.string.to-well-formed');
+
 module.exports = parent;