+++ packages/core-js/stage/3.js
@@ -3,6 +3,7 @@ var parent = require('./4');
 require('../proposals/array-from-async-stage-2');
 require('../proposals/array-grouping-stage-3-2');
 require('../proposals/change-array-by-copy');
+require('../proposals/well-formed-unicode-strings');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping-stage-3');